Move other shared components
